Moved BindingPriority to its own file.
